% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeR.R
\name{brainAge}
\alias{brainAge}
\title{brainAge}
\usage{
brainAge(x, template, model, polyOrder, batch_size = 8)
}
\arguments{
\item{x}{input image}

\item{template}{input template, optional}

\item{model}{input deep model, optional}

\item{polyOrder}{optional polynomial order for intensity matching (e.g. 1)}

\item{batch_size}{greater than 1 uses simulation to add variance in estimated values}
}
\value{
data frame of predictions and the brain age model
}
\description{
Estimate brain age and related variable from input T1 MRI
}
\examples{

\dontrun{
myPredictions = brainAge( img, template, model )
}
}
\author{
Avants BB
}
