% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeR.R
\name{brainAgePreprocessing}
\alias{brainAgePreprocessing}
\title{brainAgePreprocessing}
\usage{
brainAgePreprocessing(x, template, templateBrainMask)
}
\arguments{
\item{x}{input image}

\item{template}{input template, optional}

\item{templateBrainMask}{input template brain mask, optional}
}
\value{
preprocessing in a list
\itemize{
\item{"imageAffine": }{Affine transformed and intensity normalized image.}
\item{"brainMask": }{brain extraction probability mask.}
\item{"brainMaskAffine": }{brain extraction probability mask, affine transformed.}
\item{"biasField": }{\code{n4} bias field output}
\item{"affineMapping": }{\code{antsRegistration} output}
}
}
\description{
MRI preprocessing for brain age
}
\examples{

\dontrun{
myPre = brainAgePreprocessing( img )
}
}
\author{
Avants BB
}
