% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeR.R
\name{brainExtraction}
\alias{brainExtraction}
\title{brainExtraction}
\usage{
brainExtraction(x, template, model, batch_size = 8)
}
\arguments{
\item{x}{input image}

\item{template}{input template, optional}

\item{model}{input deep model, optional}

\item{batch_size}{greater than 1 uses simulation to add variance in estimated values}
}
\value{
brain extraction
}
\description{
ANTs brain extraction implemented with a u-net
}
\examples{

\dontrun{
myPredictions = brainExtraction( img )
}
}
\author{
Avants BB
}
