% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeR.R
\name{getBrainAgeModel}
\alias{getBrainAgeModel}
\title{getBrainAgeModel}
\usage{
getBrainAgeModel(modelPrefix)
}
\arguments{
\item{modelPrefix}{prefix identifying model file locations following \code{load_model_weights_tf}}
}
\value{
tensorflow model
}
\description{
Create the brain age model data, downloading data as necessary. Data will be
downloaded from \url{https://figshare.com/articles/pretrained_networks_for_deep_learning_applications/7246985}
}
\examples{

\dontrun{
mdl = getBrainAgeModel( tempfile() )
}
}
\author{
Avants BB
}
