% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeR.R
\name{standardizeIntensity}
\alias{standardizeIntensity}
\title{standardizeIntensity}
\usage{
standardizeIntensity(x, mask, quantiles = c(0.01, 0.99))
}
\arguments{
\item{x}{input image}

\item{mask}{input image mask}

\item{quantiles}{two-vector defining quantiles in the range of 0 to 1}
}
\value{
image
}
\description{
Robust intensity standardization method
}
\examples{

\dontrun{
imgn = standardizeIntensity( img )
}
}
\author{
Avants BB
}
