% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeR.R
\name{brainAge}
\alias{brainAge}
\title{brainAge}
\usage{
brainAge(
  x,
  template,
  templateBrainMask,
  model,
  polyOrder,
  batch_size = 8,
  sdAff = 0.01
)
}
\arguments{
\item{x}{input image}

\item{template}{input template, optional}

\item{templateBrainMask}{input template brain mask, optional}

\item{model}{input deep model, see \code{getBrainAgeModel}}

\item{polyOrder}{optional polynomial order for intensity matching (e.g. 1)}

\item{batch_size}{greater than 1 uses simulation to add variance in estimated values}

\item{sdAff}{larger values induce more variance}
}
\value{
data frame of predictions and the brain age model
}
\description{
Estimate brain age and related variable from input T1 MRI.
}
\examples{

\dontrun{
library( brainAgeR )
library( ANTsR )
library( keras )
filename = system.file("extdata", "test_image.nii.gz", package = "brainAgeR", mustWork = TRUE)
img = antsImageRead( filename ) # T1 image
mdl = getBrainAgeModel( tempfile() )
bage = brainAge( img, batch_size = 10, sdAff = 0.01, model = mdl )
bage[[1]][,1:4]
}

}
\author{
Avants BB
}
